requirements.txt
fastapi
pydantic
python-dotenv
sentence-transformers
numpy
faiss-cpu
openai
uvicorn
